+++ tests/cli/test_cli.py
@@ -1,57 +1,55 @@
-import pytest
-from unittest.mock import patch, AsyncMock
+from unittest.mock import patch, Mock
 from ytfetcher._cli import YTFetcherCLI, create_parser
 from ytfetcher.config import HTTPConfig
 from ytfetcher.services.exports import METEDATA_LIST
 
 # --> Basic Call Tests <--
 
-@pytest.mark.asyncio
+@patch('ytfetcher._cli.TXTExporter.write')
 @patch.object(YTFetcherCLI, '_run_fetcher')
-async def test_run_from_channel_called(mock_run_channel):
+def test_run_from_channel_called(mock_run_channel, mock_export_as_txt):
     parser = create_parser()
     args = parser.parse_args([
         "from_channel",
     ])
 
     cli = YTFetcherCLI(args=args)
-    await cli.run()
+    cli.run()
 
     mock_run_channel.assert_called_once()
 
-@pytest.mark.asyncio
+@patch('ytfetcher._cli.TXTExporter.write')
 @patch.object(YTFetcherCLI, '_run_fetcher')
-async def test_run_from_video_ids_called(mock_run_from_video_ids):
+def test_run_from_video_ids_called(mock_run_from_video_ids, mock_export_as_txt):
     parser = create_parser()
     args = parser.parse_args([
         "from_video_ids",
     ])
 
     cli = YTFetcherCLI(args=args)
-    await cli.run()
+    cli.run()
 
     mock_run_from_video_ids.assert_called_once()
 
-@pytest.mark.asyncio
+@patch('ytfetcher._cli.TXTExporter.write')
 @patch.object(YTFetcherCLI, '_run_fetcher')
-async def test_run_from_playlist_id_called(mock_run_from_playlist_id):
+def test_run_from_playlist_id_called(mock_run_from_playlist_id, mock_export_as_txt):
     parser = create_parser()
     args = parser.parse_args([
         "from_playlist_id",
     ])
 
     cli = YTFetcherCLI(args=args)
-    await cli.run()
+    cli.run()
 
     mock_run_from_playlist_id.assert_called_once()
 
 # --> Arguments Test <--
 
-@pytest.mark.asyncio
 @patch('ytfetcher._cli.YTFetcherCLI._initialize_http_config')
 @patch('ytfetcher._cli.YTFetcher')
-async def test_run_from_channel_arguments_passed_correctly_to_ytfetcher(mock_ytfetcher, mock_initialize_http_config):
-    mock_fetcher = AsyncMock()
+def test_run_from_channel_arguments_passed_correctly_to_ytfetcher(mock_ytfetcher, mock_export_as_txt, mock_initialize_http_config):
+    mock_fetcher = Mock()
     mock_ytfetcher.from_channel.return_value = mock_fetcher
 
     # Mock HTTP Config
@@ -64,7 +62,7 @@ async def test_run_from_channel_arguments_passed_correctly_to_ytfetcher(mock_ytf
     ])
 
     cli = YTFetcherCLI(args=args)
-    await cli.run()
+    cli.run()
 
     mock_ytfetcher.from_channel.assert_called_once_with(
         channel_handle="Channel",
@@ -76,13 +74,12 @@ async def test_run_from_channel_arguments_passed_correctly_to_ytfetcher(mock_ytf
         filters=[]
     )
 
-    mock_fetcher.fetch_youtube_data.assert_awaited_once()
+    mock_fetcher.fetch_youtube_data.assert_called_once()
 
-@pytest.mark.asyncio
 @patch('ytfetcher._cli.YTFetcherCLI._initialize_http_config')
 @patch('ytfetcher._cli.YTFetcher')
-async def test_run_from_playlist_id_arguments_passed_correctly_to_ytfetcher(mock_ytfetcher, mock_initialize_http_config):
-    mock_fetcher = AsyncMock()
+def test_run_from_playlist_id_arguments_passed_correctly_to_ytfetcher(mock_ytfetcher, mock_export_as_txt, mock_initialize_http_config):
+    mock_fetcher = Mock()
     mock_ytfetcher.from_playlist_id.return_value = mock_fetcher
 
     # Mock HTTP Config
@@ -97,7 +94,7 @@ async def test_run_from_playlist_id_arguments_passed_correctly_to_ytfetcher(mock
     ])
 
     cli = YTFetcherCLI(args=args)
-    await cli.run()
+    cli.run()
 
     mock_ytfetcher.from_playlist_id.assert_called_once_with(
         playlist_id="playlistid",
@@ -110,12 +107,11 @@ async def test_run_from_playlist_id_arguments_passed_correctly_to_ytfetcher(mock
 
     mock_fetcher.fetch_youtube_data.assert_called_once()
 
-@pytest.mark.asyncio
 @patch('ytfetcher._cli.YTFetcherCLI._initialize_http_config')
 @patch('ytfetcher._cli.TXTExporter.write')
 @patch('ytfetcher._cli.YTFetcher')
-async def test_run_from_video_ids_arguments_passed_correctly_to_ytfetcher(mock_ytfetcher, mock_export_as_txt, mock_initialize_http_config):
-    mock_fetcher = AsyncMock()
+def test_run_from_video_ids_arguments_passed_correctly_to_ytfetcher(mock_ytfetcher, mock_export_as_txt, mock_initialize_http_config):
+    mock_fetcher = Mock()
     mock_ytfetcher.from_video_ids.return_value = mock_fetcher
 
     # Mock HTTP Config
@@ -130,7 +126,7 @@ async def test_run_from_video_ids_arguments_passed_correctly_to_ytfetcher(mock_y
     ])
 
     cli = YTFetcherCLI(args=args)
-    await cli.run()
+    cli.run()
 
     mock_ytfetcher.from_video_ids.assert_called_once_with(
         video_ids=['id1', 'id2'],
@@ -145,12 +141,11 @@ async def test_run_from_video_ids_arguments_passed_correctly_to_ytfetcher(mock_y
 
 # --> Exporter Tests <--
 
-@pytest.mark.asyncio
 @patch('ytfetcher._cli.TXTExporter.write')
 @patch('ytfetcher._cli.TXTExporter')
 @patch('ytfetcher._cli.YTFetcher')
-async def test_export_method_from_video_ids(mock_ytfetcher, mock_exporter_class, mock_export_as_txt):
-    mock_fetcher = AsyncMock()
+def test_export_method_from_video_ids(mock_ytfetcher, mock_exporter_class, mock_export_as_txt):
+    mock_fetcher = Mock()
     mock_ytfetcher.from_video_ids.return_value = mock_fetcher
     mock_fetcher.fetch_youtube_data.return_value = 'channeldata'
 
@@ -166,7 +161,7 @@ async def test_export_method_from_video_ids(mock_ytfetcher, mock_exporter_class,
     ])
 
     cli = YTFetcherCLI(args=args)
-    await cli.run()
+    cli.run()
 
     mock_exporter_class.assert_called_once_with(
         channel_data='channeldata',
@@ -178,12 +173,11 @@ async def test_export_method_from_video_ids(mock_ytfetcher, mock_exporter_class,
 
     mock_exporter_instance.write.assert_called_once()
 
-@pytest.mark.asyncio
 @patch('ytfetcher._cli.TXTExporter.write')
 @patch('ytfetcher._cli.TXTExporter')
 @patch('ytfetcher._cli.YTFetcher')
-async def test_export_method_from_channel(mock_ytfetcher, mock_exporter_class, mock_export_as_txt):
-    mock_fetcher = AsyncMock()
+def test_export_method_from_channel(mock_ytfetcher, mock_exporter_class, mock_export_as_txt):
+    mock_fetcher = Mock()
     mock_ytfetcher.from_channel.return_value = mock_fetcher
     mock_fetcher.fetch_youtube_data.return_value = 'channeldata'
 
@@ -197,7 +191,7 @@ async def test_export_method_from_channel(mock_ytfetcher, mock_exporter_class, m
     ])
 
     cli = YTFetcherCLI(args=args)
-    await cli.run()
+    cli.run()
 
     mock_exporter_class.assert_called_once_with(
         channel_data='channeldata',
@@ -209,12 +203,11 @@ async def test_export_method_from_channel(mock_ytfetcher, mock_exporter_class, m
 
     mock_exporter_instance.write.assert_called_once()
 
-@pytest.mark.asyncio
 @patch('ytfetcher._cli.TXTExporter.write')
 @patch('ytfetcher._cli.TXTExporter')
 @patch('ytfetcher._cli.YTFetcher')
-async def test_export_method_from_playlist_id(mock_ytfetcher, mock_exporter_class, mock_export_as_txt):
-    mock_fetcher = AsyncMock()
+def test_export_method_from_playlist_id(mock_ytfetcher, mock_exporter_class, mock_export_as_txt):
+    mock_fetcher = Mock()
     mock_ytfetcher.from_playlist_id.return_value = mock_fetcher
     mock_fetcher.fetch_youtube_data.return_value = 'channeldata'
 
@@ -228,7 +221,7 @@ async def test_export_method_from_playlist_id(mock_ytfetcher, mock_exporter_clas
     ])
 
     cli = YTFetcherCLI(args=args)
-    await cli.run()
+    cli.run()
 
     mock_exporter_class.assert_called_once_with(
         channel_data='channeldata',